commit 4

+++ js/signup.js
@@ -4,11 +4,14 @@ window.addEventListener('load',function(){
     const emailInput = document.getElementById('email');
     const passwordInput = document.getElementById('password');
     const confirmPasswordInput = document.getElementById('confirmPassword');
+    const userTypeInput = document.getElementById("userType");
+
 
     const usernameError = document.getElementById('usernameError');
     const emailError = document.getElementById('emailError');
     const passwordError = document.getElementById('passwordError');
     const confirmError = document.getElementById('confirmError');
+    const userTypeError = document.getElementById('userTypeError');
 
     const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
     const passwordRegex = /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,}$/;
@@ -50,11 +53,26 @@ window.addEventListener('load',function(){
     return true;
     }
 
+    ///////////////////
+    function validateUserType() {
+    if (userTypeInput.value === "")
+    {
+            userTypeError.textContent='Please select a user type';
+        return false;
+      }
+      userTypeError.textContent='';
+      return true;
+    }
+    ////////////////
+
 
     usernameInput.addEventListener('blur', validateUsername);
     emailInput.addEventListener('blur', validateEmail);
     passwordInput.addEventListener('blur', validatePassword);
     confirmPasswordInput.addEventListener('blur', validateConfirmPassword);
+//////////
+    userTypeInput.addEventListener('blur', validateUserType);
+/////////////
 
 
     document.getElementById('signupForm').addEventListener('submit', function(e) {
@@ -64,14 +82,20 @@ window.addEventListener('load',function(){
     const isEmailValid = validateEmail();
     const isPasswordValid = validatePassword();
     const isConfirmValid = validateConfirmPassword();
+    ////////
+    const isUserTypeValid = validateUserType();
+    /////////
 
-
-    if (isUsernameValid && isEmailValid && isPasswordValid && isConfirmValid) {
-
+//////
+    if (isUsernameValid && isEmailValid && isPasswordValid && isConfirmValid && isUserTypeValid) {
+//////
         const userData = {
         username: usernameInput.value.trim(),
         email: emailInput.value.trim(),
-        password: passwordInput.value.trim()
+        password: passwordInput.value.trim(),
+        ////
+        userType: userTypeInput.value
+        ////
         };
     
         localStorage.setItem('animeUser', JSON.stringify(userData));
